+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.10 (unreleased)
+## 0.0.10 (2018-06-13)
 
 -   Model the time extent of events and superevents using the
     `glue.segments` module.